requirements.txt
streamlit==0.65.2
numpy==1.19.1
opencv-python==4.4.0.44
Pillow==7.2.0
opencv-python-headless==4.2.0.32
h5py==2.10.0
tensorflow-cpu==2.2.0
